Move `InstantInterval` to main library; update Kotlin 2.3.10, datetime 0.7.1
Starting from Kotlin 2.1.20, `Instant` from kotlinx.datetime was introduced in the Kotlin base library. It has been removed in kotlinx.datetime 0.7.0.
