+++ build-logic/build.gradle.kts
@@ -8,6 +8,6 @@ repositories {
 }
 
 dependencies {
-    implementation("org.jetbrains.kotlin:kotlin-gradle-plugin:2.1.21")
+    implementation("org.jetbrains.kotlin:kotlin-gradle-plugin:2.3.10")
     implementation("org.jetbrains.dokka:dokka-gradle-plugin:2.0.0")
 }
