+++ kotlinx.interval/src/commonMain/kotlin/InstantInterval.kt
@@ -1,10 +1,10 @@
-package io.github.whathecode.kotlinx.interval.datetime
+package io.github.whathecode.kotlinx.interval
 
 import io.github.whathecode.kotlinx.interval.Interval
 import io.github.whathecode.kotlinx.interval.IntervalTypeOperations
-import kotlinx.datetime.Instant
 import kotlin.math.absoluteValue
 import kotlin.time.Duration
+import kotlin.time.Instant
 
 
 /**
